requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-default-group==1.2.4
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
packaging==25.0
pluggy==1.6.0
psycopg[binary]==3.2.1
pytest==8.3.3
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
